new script: urlcleaner
This script takes an URL from the clipboard, removes tracking parameters
(everything after the question mark) and copies the cleaned-up URL back
to the clipboard.

Alternatively, if the script is called with the parameter "go", it will
open the cleaned-up URL in Firefox instead of copying it to the
clipboard.